Zone added Random Block Generator reduced to only one and non-relative instance Health attack value reduced to 0.1

+++ Assets/Scripts/Player.cs
@@ -48,7 +48,7 @@ public class Player : MonoBehaviour
             health += addValue;
             if (healthMeter)
             {
-                healthMeter.text = "❤ " + health;
+                healthMeter.text = "❤ " + (int)health;
             }
         }
         if (health < 1)
